Move util function to live with other date functions
