+++ src/sentry/static/sentry/app/utils/dates.jsx
@@ -166,3 +166,27 @@ export function intervalToMilliseconds(interval) {
   };
   return parseInt(value, 10) * multipliers[unit] * 1000;
 }
+
+
+/**
+ * This parses our period shorthand strings (e.g. <int><unit>)
+ * and converts it into hours
+ */
+export function parsePeriodToHours(str) {
+  const [, periodNumber, periodLength] = str.match(/([0-9]+)([smhdw])/);
+
+  switch (periodLength) {
+    case 's':
+      return periodNumber / (60 * 60);
+    case 'm':
+      return periodNumber / 60;
+    case 'h':
+      return periodNumber;
+    case 'd':
+      return periodNumber * 24;
+    case 'w':
+      return periodNumber * 24 * 7;
+    default:
+      return -1;
+  }
+}
